refactor(convert): extract pseudo module
Phase 1 of fulgur-odgo. Move 13 pseudo-related fns from convert/mod.rs
into convert/pseudo.rs: build_pseudo_image, is_block_pseudo,
wrap_with_pseudo_content, node_has_block_pseudo_image,
node_has_inline_pseudo_image, node_has_absolute_pseudo,
build_block_pseudo_images, wrap_with_block_pseudo_images,
build_inline_pseudo_image, attach_link_to_inline_image,
inject_inline_pseudo_images, effective_pseudo_size_px (all pub(super))
plus resolve_pseudo_size (private, only used internally by
build_pseudo_image).

is_pseudo_node stays in mod.rs as a shared helper used by both the
dispatcher and (in later tasks) positioned.rs.

No behavioral change. VRT byte-identical against committed goldens.
